Updated table
