+++ client/src/App.tsx
@@ -3,6 +3,7 @@ import './css/App.css';
 // Components
 import {Table} from './components/Table'
 import {Login} from './components/Login'
+import { BudgetView } from './components/BudgetView';
 
 function App() {
   const [loggedIn, setLoggedIn] = React.useState(false)
@@ -15,7 +16,7 @@ function App() {
   return (
     <div className="App">
       {!loggedIn && <Login toggleLogin={toggleLogin}/>}
-      {loggedIn && [<Table/>, <button onClick={toggleLogin}>Logout</button>]}
+      {loggedIn && [<BudgetView/>, <button onClick={toggleLogin}>Logout</button>]}
     </div>
   );
 }
